+++ src/main/java/com/eventos/apoio/repository/SessaoRepository.java
@@ -3,5 +3,5 @@ package com.eventos.apoio.repository;
 import com.eventos.apoio.model.Sessao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SessaoRepository extends JpaRepository<Sessao, Long> {
+public interface SessaoRepository extends JpaRepository<Sessao, Integer> {
 }
